feat: Share database mongodb
