+++ packages/frontend/editor-ui/src/components/Node/NodeCreator/composables/useActionsGeneration.ts
@@ -20,7 +20,7 @@ import { i18n } from '@n8n/i18n';
 import { getCredentialOnlyNodeType } from '@/utils/credentialOnlyNodes';
 import { formatTriggerActionName } from '../utils';
 import { useEvaluationStore } from '@/stores/evaluation.store.ee';
-import { filterNodeOperationsByUserRole } from '@/utils/rbacUtils';
+import { filterNodeOperationsByUserRole } from '@/utils/nodeOperationFilters';
 
 const PLACEHOLDER_RECOMMENDED_ACTION_KEY = 'placeholder_recommended';
